\name{Copula.surv-package}
\alias{Copula.surv-package}
\alias{Copula.surv}
\docType{package}
\title{
Association Analysis of Bivariate Survival Data Based on Copulas
}
\description{
 Perform association analysis of bivariate survival data based on copula models.
 Two different ways to estimate the association parameter in copula models are implemented.
 A goodness-of-fit test for a given copula model is implemented.
 See Emura, Lin and Wang (2010) <doi:10.1016/j.csda.2010.03.013> for details.
 A more general reference for the analysis of bivariate survival data is Emura, Matsui, and Rondeau (2019).
}
\details{
Details are seen from the references.
}
\author{
Takeshi Emura
Maintainer: Takeshi Emura <takeshiemura@gmail.com>
}
\references{
Emura T, Lin CW, Wang W (2010) A goodness-of-fit test for Archimedean copula models
 in the presence of right censoring, Compt Stat Data Anal 54: 3033-43

Emura T, Matsui S, Rondeau V (2019), Survival Analysis with Correlated Endpoints, Joint Frailty-Copula Models, JSS Research Series in Statistics, Springer
}
\keyword{ package }
