\name{Copula.surv-package}
\alias{Copula.surv-package}
\alias{Copula.surv}
\docType{package}
\title{
Analysis of Doubly-Truncated Data
}
\description{
 Perform association analysis of bivariate survival data based on copula models.
 Two different ways to estimate the association parameter in copula models are implemented.
 A goodness-of-fit test for a given copula model is implemented.
 See Emura, Lin and Wang (2010) <doi.org/10.1016/j.csda.2010.03.013> for details.
}
\details{
Details are seen from the references.
}
\author{
Takeshi Emura
Maintainer: Takeshi Emura <takeshiemura@gmail.com>
}
\references{
Emura T, Lin CW, Wang W (2010) A goodness-of-fit test for Archimedean copula models
 in the presence of right censoring, Compt Stat Data Anal 54: 3033-43
}
\keyword{ package }